+++ resources/V7_db_speed_up.sql
@@ -1,0 +1,5 @@
+create index idx_dao_popularity_idx
+    on daos (popularity_index desc);
+
+create index idx_dao_category
+    on daos (categories);