+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Indexes for slow queries
+
 ## [0.1.8] - 2024-03-22
 
 ### Added